+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 # 2.14.0 (2022-11-05)
 
 * Add device and author information as tags [Issue 52](https://github.com/grasshopper7/extentreports-cucumber6-adapter/issues/52)
+* Base folder report name custom delimiter setting [Issue 53](https://github.com/grasshopper7/extentreports-cucumber6-adapter/issues/53)
 
 # 2.13.1 (2022-11-02)